+++ GiganticEmu.Agent/Pages/Error.cshtml.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics;
+using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
@@ -11,7 +12,9 @@ namespace GiganticEmu.Agent.Pages
     {
         public string? RequestId { get; set; }
 
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+        public string? ErrorMessage { get; set; }
+
+        public string? StackTrace { get; set; }
 
         private readonly ILogger<ErrorModel> _logger;
 
@@ -23,6 +26,16 @@ namespace GiganticEmu.Agent.Pages
         public void OnGet()
         {
             RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier;
+
+            var exceptionHandlerPathFeature = HttpContext.Features.Get<IExceptionHandlerPathFeature>();
+
+            ErrorMessage = exceptionHandlerPathFeature?.Error?.Message;
+            StackTrace = exceptionHandlerPathFeature?.Error?.StackTrace;
+        }
+
+        public void OnPost()
+        {
+            OnGet();
         }
     }
 }